Switched ints to BigIntegers so no overflow errors for huge numbers, implemented calling by users for all number list types

+++ Mathmagician/FibonacciNumbers.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Numerics;
 using System.Threading.Tasks;
 
 namespace Mathmagician
@@ -14,9 +15,9 @@ namespace Mathmagician
             increment = 0;
         }
 
-        public override int GetNext(int previous_number)
+        public override BigInteger GetNext(BigInteger previous_number)
         {
-            int next_number = base.GetNext(previous_number);
+            BigInteger next_number = base.GetNext(previous_number);
             increment = previous_number;
             return next_number;
         }
